[backend/frontend] Add fintel to StixDomainObject (#10962)
