+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectContent.jsx
@@ -26,6 +26,7 @@ import withRouter from '../../../../utils/compat_router/withRouter';
 import CKEditor from '../../../../components/CKEditor';
 import { htmlToPdf } from '../../../../utils/htmlToPdf/htmlToPdf';
 import HtmlDisplay from '../../../../components/HtmlDisplay';
+import { typesWithFintelTemplates } from '../../../../utils/hooks/useAttributes';
 
 pdfjs.GlobalWorkerOptions.workerSrc = `${APP_BASE_PATH}/static/ext/pdf.worker.mjs`;
 
@@ -486,6 +487,7 @@ class StixCoreObjectContentComponent extends Component {
     const { innerHeight } = window;
     const height = innerHeight - 320;
     const isContentCompatible = isContainerWithContent(stixCoreObject.entity_type);
+    const hasFintelTemplates = typesWithFintelTemplates.includes(stixCoreObject.entity_type);
     return (
       <div className={classes.container} data-testid="sco-content-page">
         <StixCoreObjectContentFiles
@@ -501,7 +503,7 @@ class StixCoreObjectContentComponent extends Component {
           currentFileId={currentFileId}
           onFileChange={this.handleFileChange.bind(this)}
           filesFromTemplate={filesFromTemplate}
-          hasOutcomesTemplate={isContentCompatible}
+          hasOutcomesTemplate={isContentCompatible || hasFintelTemplates}
         />
         {isLoading ? (
           <Loader variant={LoaderVariant.inElement} />
@@ -851,7 +853,7 @@ const StixCoreObjectContent = createRefetchContainer(
             }
           }
         }
-        ... on Container {
+        ... on StixDomainObject {
           filesFromTemplate(first: 500) @connection(key: "Pagination_filesFromTemplate") {
             edges {
               node {
