+++ opencti-platform/opencti-graphql/src/modules/channel/channel.graphql
@@ -100,6 +100,16 @@ type Channel implements BasicObject & StixObject & StixCoreObject & StixDomainOb
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Channel
     name: String!
     description: String
