+++ opencti-platform/opencti-graphql/src/modules/dataComponent/dataComponent.graphql
@@ -101,6 +101,16 @@ type DataComponent implements BasicObject & StixObject & StixCoreObject & StixDo
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Technical
     creators: [Creator!]
     toStix(version: Version): String
