+++ opencti-platform/opencti-graphql/src/modules/dataSource/dataSource.graphql
@@ -101,6 +101,16 @@ type DataSource implements BasicObject & StixObject & StixCoreObject & StixDomai
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Technical
     creators: [Creator!]
     toStix(version: Version): String
