+++ opencti-platform/opencti-graphql/src/modules/entitySetting/entitySetting-utils.ts
@@ -75,7 +75,7 @@ export const defaultScale = JSON.stringify({
 
 // Available settings works by override.
 export const availableSettings: Record<string, Array<string>> = {
-  [ABSTRACT_STIX_DOMAIN_OBJECT]: ['attributes_configuration', 'platform_entity_files_ref', 'platform_hidden_type', 'enforce_reference', 'workflow_configuration'],
+  [ABSTRACT_STIX_DOMAIN_OBJECT]: ['attributes_configuration', 'platform_entity_files_ref', 'platform_hidden_type', 'enforce_reference', 'workflow_configuration', 'templates'],
   [ABSTRACT_STIX_CORE_RELATIONSHIP]: ['attributes_configuration', 'enforce_reference', 'workflow_configuration'],
   [STIX_SIGHTING_RELATIONSHIP]: ['attributes_configuration', 'enforce_reference', 'platform_hidden_type', 'workflow_configuration'],
   [ABSTRACT_STIX_CYBER_OBSERVABLE]: ['platform_hidden_type'],
