+++ opencti-platform/opencti-graphql/src/modules/event/event.graphql
@@ -100,6 +100,16 @@ type Event implements BasicObject & StixCoreObject & StixDomainObject & StixObje
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Event
     name: String!
     description: String
