+++ opencti-platform/opencti-graphql/src/modules/language/language.graphql
@@ -100,6 +100,16 @@ type Language implements BasicObject & StixCoreObject & StixDomainObject & StixO
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Language
     name: String!
     aliases: [String]
