+++ opencti-platform/opencti-graphql/src/modules/malwareAnalysis/malwareAnalysis.graphql
@@ -100,6 +100,16 @@ type MalwareAnalysis implements BasicObject & StixCoreObject & StixDomainObject 
     objectAssignee: [Assignee!]
     objectParticipant: [Participant!]
     avatar: OpenCtiFile
+    fintelTemplates: [FintelTemplate!]
+    filesFromTemplate(
+        first: Int
+        prefixMimeType: String
+        after: ID
+        orderBy: FileOrdering
+        orderMode: OrderingMode
+        search: String
+        filters: FilterGroup
+    ): FileConnection
     # Technical
     creators: [Creator!]
     toStix(version: Version): String
