+++ opencti-platform/opencti-graphql/src/modules/securityPlatform/securityPlatform.graphql
@@ -100,6 +100,16 @@ type SecurityPlatform implements BasicObject & StixObject & StixCoreObject & Sti
   objectAssignee: [Assignee!]
   objectParticipant: [Participant!]
   avatar: OpenCtiFile
+  fintelTemplates: [FintelTemplate!]
+  filesFromTemplate(
+    first: Int
+    prefixMimeType: String
+    after: ID
+    orderBy: FileOrdering
+    orderMode: OrderingMode
+    search: String
+    filters: FilterGroup
+  ): FileConnection
   # Identity
   identity_class: String!
   name: String!
