+++ opencti-platform/opencti-graphql/src/modules/threatActorIndividual/threatActorIndividual.graphql
@@ -110,6 +110,16 @@ type ThreatActorIndividual implements BasicObject & StixObject & StixCoreObject 
   objectAssignee: [Assignee!]
   objectParticipant: [Participant!]
   avatar: OpenCtiFile
+  fintelTemplates: [FintelTemplate!]
+  filesFromTemplate(
+    first: Int
+    prefixMimeType: String
+    after: ID
+    orderBy: FileOrdering
+    orderMode: OrderingMode
+    search: String
+    filters: FilterGroup
+  ): FileConnection
   # ThreatActorIndividual
   name: String!
   description: String
